UI Changes. Drop shadow jumobotron and green pastel modals

+++ TDC/TDC/Tools/UserActions.cs
@@ -132,20 +132,23 @@ namespace TDC.Tools
 
         public static string messageParser(User user)
         {
-            ApplicationDbContext db = new ApplicationDbContext(); 
-            
             string final = "";
-            var list =  user.Message.ToList();
-            foreach (var item in user.Message)
+            if (user != null)
             {
-                final += item.notification + "\n\n";
+                ApplicationDbContext db = new ApplicationDbContext();
 
-                Message delete = db.Message.Find(item.ID);
-                db.Message.Remove(delete);
-                db.SaveChanges();
                 
-            }   
-            
+                var list = user.Message.ToList();
+                foreach (var item in user.Message)
+                {
+                    final += item.notification + "\n\n";
+
+                    Message delete = db.Message.Find(item.ID);
+                    db.Message.Remove(delete);
+                    db.SaveChanges();
+
+                }
+            }
             return final;
